refactor: 강의 추천/리뷰/사용자 컴포넌트 구조 및 네이밍 변경, 불필요 파일 삭제(#219)
- LoggedInCourses를 LoggedIn으로 이름 변경 및 user-course 디렉토리로 이동
- UserCourseSection을 UserCourse로 이름 변경 및 user-course 디렉토리로 이동
- ReviewSection 컴포넌트 경로를 course-content로 이동
- RecommendedSection 컴포넌트 파일 삭제
- 강의 추천/리뷰/사용자 관련 컴포넌트 구조 및 네이밍 일관성 개선
